// components/root/ModalsTenant/forms.interfaces.ts
export interface IFormOne {
    nit: string
    businessName: string
    address: string
    responsibleId: string
    responsibleName: string
    responsibleLastName: string
    responsiblePhone: string
    email: string
}

export interface IFormTwo {
    subdomain: string
}

export interface IFormThree {
    username: string
    name: string
    lastName: string
    email: string
    password1: string
    password2: string
}